Update 20-dns.conflist
Suggesting that people just fill-in the last bit of a legitimately-started, fake Ubiquiti mac address.  Very unlikely anyone will have an Iridium satellite on their network, thus I've used the prefix for Iridium here :) 00:1C:B4.   Advantage is that people will now start this with a legit prefix and know the FORMAT to prevent dashes / cruft.