fix(nspawn-container): Retry 0-setup-system.sh on failure
Other system processes can acquire a dpkg frontend lock on startup that
will cause the downloaded package install to fail. The setup script
continues without exiting and machinectl is not found.

This change forces the script to exit early on error, and adds 5 retries
to the systemd unit at 30s intervals.

Signed-off-by: Micah Hausler <hausler.m@gmail.com>